Add option variable for setting the main elm file
This change adds an option-var for setting a main .elm file that will
always be compiled when flycheck-elm checks any file.

This is helpful because the output of elm-make compilation is only
usable if the .elm file has a 'main' function, so when flycheck-elm
checks and compiles any non-main files, the resulting elm.js or
index.html file isn't usable until the main elm file is compiled again.