Use flycheck's new :default-directory option to handle projects with subdirs
This enables the checker if and only if an existing elm-package.json can
be found, and runs elm-make from the directory of that file.

The previous behaviour of allowing elm-make to implicitly create
elm-package.json in the current directory - which could cause the
checker initialisation to hang - is also therefore avoided.

Fixes #3.